requirements.txt
adtk==0.6.2
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
awscli==1.32.84
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.1
azure-mgmt-authorization==4.0.0
azure-mgmt-core==1.4.0
azure-mgmt-loganalytics==12.0.0
azure-mgmt-resource==23.1.0
azure-monitor-query==1.3.0
backoff==2.2.1
bcrypt==4.1.3
billiard==3.6.4.0
boto3==1.34.84
botocore==1.34.84
cachetools==5.3.3
celery==5.2.7
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickhouse-connect==0.7.8
colorama==0.4.4
contourpy==1.2.1
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.5
cx_Oracle==8.3.0
cycler==0.12.1
datadog-api-client==2.24.0
defusedxml==0.7.1
distro==1.9.0
dj-email-url==1.0.6
dj-rest-auth==2.2.7
Django==4.1.13
django-allauth==0.52.0
django-appconf==1.0.6
django-celery-beat==2.4.0
django-celery-email==3.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-prometheus==2.4.0.dev5
django-redis==5.2.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.16
elastic-transport==8.13.1
elasticsearch==8.14.0
environs==11.0.0
exceptiongroup==1.2.0
fdb==2.0.2
filelock==3.14.0
fonttools==4.51.0
future==1.0.0
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gql==3.5.0
graphql-core==3.2.3
grpcio==1.62.0
grpcio-status==1.62.0
grpcio-tools==1.51.1
gunicorn==22.0.0
h11==0.14.0
hdbcli==2.20.23
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
isodate==0.6.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.0
kaleido==0.2.1
kiwisolver==1.4.5
kombu==5.3.7
kubernetes==29.0.0
lz4==4.3.3
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.5
mypy-protobuf==3.4.0
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.25.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
patsy==0.5.6
pdpyras==5.2.0
pg8000==1.31.2
pillow==10.3.0
platformdirs==4.2.2
plotly==5.21.0
portalocker==2.8.2
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.21.12
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyodbc==5.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==4.5.4
requests==2.32.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.7.2
s3transfer==0.10.1
scikit-learn==1.4.2
scipy==1.13.0
scramp==1.4.5
simplejson==3.19.2
six==1.16.0
slack_sdk==3.27.1
sniffio==1.3.1
snowflake-connector-python==3.10.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
sqlparse==0.5.0
statsmodels==0.14.2
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.5.0
tomlkit==0.12.5
tqdm==4.66.2
trino==0.328.0
trio==0.25.0
trio-websocket==0.11.1
types-protobuf==4.21.0.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.7.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
zstandard==0.22.0
